Revert "Refactor admin.go into 13 focused modules (#282) (#406)"
This reverts commit 912333de38c0526c38b0a6ff567f25bbb8cc69dc.
