--- internal/coord/admin_panels_test.go
@@ -1,47 +1,0 @@
-package coord
-
-import (
-	"net/http"
-	"net/http/httptest"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-)
-
-func TestValidatePluginURL(t *testing.T) {
-	tests := []struct {
-		name    string
-		url     string
-		wantErr bool
-	}{
-		{"valid https", "https://example.com/plugin.js", false},
-		{"valid http", "http://localhost:3000/plugin.js", false},
-		{"javascript scheme", "javascript:alert(1)", true},
-		{"data scheme", "data:text/html,<h1>Hi</h1>", true},
-		{"file scheme", "file:///etc/passwd", true},
-		{"no scheme", "example.com/plugin.js", true},
-		{"no host", "https://", true},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := validatePluginURL(tt.url)
-			if tt.wantErr {
-				require.Error(t, err)
-			} else {
-				require.NoError(t, err)
-			}
-		})
-	}
-}
-
-func TestHandleSystemHealth_MethodNotAllowed(t *testing.T) {
-	srv := newTestServerWithS3AndBucket(t)
-
-	req := httptest.NewRequest(http.MethodPost, "/api/system/health", nil)
-	rec := httptest.NewRecorder()
-	srv.adminMux.ServeHTTP(rec, req)
-
-	assert.Equal(t, http.StatusMethodNotAllowed, rec.Code)
-}
