--- internal/coord/admin_wireguard.go
@@ -1,152 +1,0 @@
-package coord
-
-import (
-	"encoding/json"
-	"io"
-	"net/http"
-	"strings"
-	"time"
-)
-
-// handleWGClients handles GET (list) and POST (create) for WireGuard clients.
-func (s *Server) handleWGClients(w http.ResponseWriter, r *http.Request) {
-	switch r.Method {
-	case http.MethodGet:
-		s.handleWGClientsList(w, r)
-	case http.MethodPost:
-		s.handleWGClientCreate(w, r)
-	default:
-		s.jsonError(w, "method not allowed", http.StatusMethodNotAllowed)
-	}
-}
-
-// handleWGClientsList returns all WireGuard clients by proxying to the concentrator.
-func (s *Server) handleWGClientsList(w http.ResponseWriter, r *http.Request) {
-	// Proxy to concentrator via relay
-	respBody, err := s.relay.SendAPIRequest(r.Context(), "GET /clients", nil, 10*time.Second)
-	if err != nil {
-		s.jsonError(w, "concentrator not available: "+err.Error(), http.StatusServiceUnavailable)
-		return
-	}
-
-	// Parse API response
-	var apiResp struct {
-		StatusCode int             `json:"status_code"`
-		Body       json.RawMessage `json:"body"`
-	}
-	if err := json.Unmarshal(respBody, &apiResp); err != nil {
-		s.jsonError(w, "invalid response from concentrator", http.StatusBadGateway)
-		return
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	if apiResp.StatusCode != 200 {
-		w.WriteHeader(apiResp.StatusCode)
-	}
-	_, _ = w.Write(apiResp.Body)
-}
-
-// handleWGClientCreate creates a new WireGuard client by proxying to the concentrator.
-func (s *Server) handleWGClientCreate(w http.ResponseWriter, r *http.Request) {
-	// Read request body
-	body, err := io.ReadAll(r.Body)
-	if err != nil {
-		s.jsonError(w, "failed to read request body", http.StatusBadRequest)
-		return
-	}
-
-	// Proxy to concentrator via relay
-	respBody, err := s.relay.SendAPIRequest(r.Context(), "POST /clients", body, 10*time.Second)
-	if err != nil {
-		s.jsonError(w, "concentrator not available: "+err.Error(), http.StatusServiceUnavailable)
-		return
-	}
-
-	// Parse API response
-	var apiResp struct {
-		StatusCode int             `json:"status_code"`
-		Body       json.RawMessage `json:"body"`
-	}
-	if err := json.Unmarshal(respBody, &apiResp); err != nil {
-		s.jsonError(w, "invalid response from concentrator", http.StatusBadGateway)
-		return
-	}
-
-	// Try to extract client info for DNS cache update
-	if apiResp.StatusCode == 201 {
-		var createResp struct {
-			Client struct {
-				DNSName string `json:"dns_name"`
-				MeshIP  string `json:"mesh_ip"`
-			} `json:"client"`
-		}
-		if err := json.Unmarshal(apiResp.Body, &createResp); err == nil && createResp.Client.DNSName != "" {
-			s.peersMu.Lock()
-			s.dnsCache[createResp.Client.DNSName] = createResp.Client.MeshIP
-			s.peersMu.Unlock()
-		}
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(apiResp.StatusCode)
-	_, _ = w.Write(apiResp.Body)
-}
-
-// handleWGClientByID handles GET, PATCH, DELETE for a specific WireGuard client by proxying to the concentrator.
-func (s *Server) handleWGClientByID(w http.ResponseWriter, r *http.Request) {
-	// Extract client ID from path (admin served at /api/wireguard/clients/)
-	id := strings.TrimPrefix(r.URL.Path, "/api/wireguard/clients/")
-	if id == "" {
-		s.jsonError(w, "client ID required", http.StatusBadRequest)
-		return
-	}
-
-	var method string
-	var body []byte
-	var err error
-
-	switch r.Method {
-	case http.MethodGet:
-		method = "GET /clients/" + id
-
-	case http.MethodPatch:
-		method = "PATCH /clients/" + id
-		body, err = io.ReadAll(r.Body)
-		if err != nil {
-			s.jsonError(w, "failed to read request body", http.StatusBadRequest)
-			return
-		}
-
-	case http.MethodDelete:
-		method = "DELETE /clients/" + id
-
-	default:
-		s.jsonError(w, "method not allowed", http.StatusMethodNotAllowed)
-		return
-	}
-
-	// Proxy to concentrator via relay
-	respBody, err := s.relay.SendAPIRequest(r.Context(), method, body, 10*time.Second)
-	if err != nil {
-		s.jsonError(w, "concentrator not available: "+err.Error(), http.StatusServiceUnavailable)
-		return
-	}
-
-	// Parse API response
-	var apiResp struct {
-		StatusCode int             `json:"status_code"`
-		Body       json.RawMessage `json:"body"`
-	}
-	if err := json.Unmarshal(respBody, &apiResp); err != nil {
-		s.jsonError(w, "invalid response from concentrator", http.StatusBadGateway)
-		return
-	}
-
-	// Note: DNS cache cleanup for deleted clients is handled by periodic DNS sync
-
-	w.Header().Set("Content-Type", "application/json")
-	if apiResp.StatusCode != 200 {
-		w.WriteHeader(apiResp.StatusCode)
-	}
-	_, _ = w.Write(apiResp.Body)
-}
